segregrated codes into 3 files, main.py, sections.py, questions.py
